Revise: image changed in Partnership intro page.
